Cleaning & Commenting

+++ scripts/primaryLoad.js
@@ -698,43 +698,56 @@ $(document).ready(() => {
 	markerPlacement();
 	// Add all roulette table listeners.
 	addListeners(betArrays());
+	// Introduce a small delay to allow the tooltips and modals to properly initialize.
 	setTimeout(() => {
 		// Initialize the modals and tooltips after a small delay.
 	    var tooltipElems = document.querySelectorAll(".tooltipped"),
 	    	modalElems = document.querySelectorAll(".modal"),
 	    	tooltipInstances = M.Tooltip.init(tooltipElems),
     		modalInstances = M.Modal.init(modalElems);
+    	// Check if a parameters file exists.
 		if(!fs.existsSync(path.join(__dirname, "../parameters.json"))) {
+			// Iterate through all modal instances until the helpModal is found.
 			for(var k = 0; k < modalInstances.length; k++) {
 				if(modalInstances[k].id == "helpModal") {
+					// Change the title of the helpModal from "Help" to "Introduction" on an initial load.
 					modalInstances[k].el.children[0].children[0].innerText = "Introduction";
+					// Open the introductory modal.
 					modalInstances[k].open();
+					// Define the callback that will handle a change in the helpModal title.
 					var mutationCaller = (mutationsList, observer) => {
 					    mutationsList.forEach(mutation => {
-					        if (mutation.attributeName === 'class') {
+					        if (mutation.attributeName === "class") {
 					            if(mutation.target.id == "helpModal" && !mutation.target.classList.contains("open")) {
 									mutation.target.children[0].children[0].innerText = "Help";
 					            }
 					        }
 					    })
 					};
+					// Define an observer and let it trigger an event whenever the class list of the helpModal changes.
 					const mutationObserver = new MutationObserver(callback);
 					mutationObserver.observe(modalInstances[k].el, { attributes: true });
 				}
 			}
 		}
 		else {
+			// If a parameters file exists then check the introductionCheck on the helpModal.
 			const parameters = JSON.parse(fs.readFileSync(path.join(__dirname, "../parameters.json"), "UTF8"));
 			if(parameters.introduction == false) {
-				document.getElementById("introductionCheck").checked = true;		
+				document.getElementById("introductionCheck").checked = true;
 			}
 		}
+		// Listen for a check/uncheck on the introductionCheck located in the helpModal.
 		$("#introductionCheck").change(() => {
+			// If the checkbox is checked then write a parameters file to indicate that the introductory message is not to be showed.
 			if(document.getElementById("introductionCheck").checked) {
-				fs.writeFileSync(path.join(__dirname, "../parameters.json"), JSON.stringify({"introduction": false}), "UTF8");	
+				fs.writeFileSync(path.join(__dirname, "../parameters.json"), JSON.stringify({"introduction": false}), "UTF8");
 			}
+			// Otherwise delete the parameters file.
 			else {
-				fs.unlinkSync(path.join(__dirname, "../parameters.json"));
+				if(fs.existsSync(path.join(__dirname, "../parameters.json"))) {
+					fs.unlinkSync(path.join(__dirname, "../parameters.json"));
+				}
 			}
 		});
 	}, 50);
